Push down state management of insights section related things
